[Service] Adds a NewsService
The issue:
We want to network with the News service which has an endpoint. This layer does the conversion of Data in [Article] in its fetchArticles method. In the future, this class can be expanded and more methods can be added.

This CR:
Adds a NewsService which consumes the Networking features that were added earlier in the Networking Swift package