[ViewModel] Adds a thread-safe NewsStore that acts as in-memory cache
The issue:
We want to be able to store the news article. This class currently doesn’t provide full-fledged caching functionality but in the future this will be extended to provide in-memory and on-disc caching. One reason we can’t do this right now is because the API doesn’t support paging. Also, it doesn’t provide a way to uniquely identify a response and a TTL value of the response. When these are available, we can use these to implement a cache based on the TTL, page, and the unique identifier.

This CR:
Adds a thread-safe news store