[Refactor] Move utility extension methods into a Swift package
The issue:
The extension methods on Strnig and UserDefaults are reusable and hence can be put into a Swift package for reuse in other projcets in the future. Later, we can publish this Swift package as a public package on Github.

This CR:
Creates a Swift package to house nifty utility methods