[Fix] Clear NewStore before refresh data
The issue:
When we refresh data, the NewsStore should be cleared to make room for the new data. Later when we support proper caching, this could be optimized by checking the TTL on the cache and if it hasn’t expired yet, we can just returnthe cached data instead of making a new request.

This CR:
Accomodates clearing of NewsStore and does it before we make a fresh request to service.