[ViewModel] Add English to Martian translation logic
The issue:
Per requirement, a word can be translated if it has:
* more than 3 characters
* Is not a number unless the number is withing a String
* Maintain capitalization
* Also, account for trailing punctuation to preserve them post translation

We use regex to to find if the word has any character [a-zA-Z] and if it has then we check if it has more than 3 characters. Also, we take into account the capitalization. If it doesn’t match the regex, then it means it is a number. Numer could have CharacterSet.symbol in it and it could have .(period) or other special symbols and they are not translated as the number in that case is not considered to be within a word but rather considered to denote a decimal, a bullt point(1.2.3.4) or a curreny value ($12,000) to name a few examples.

This CR:
Adds business logic for English -> Martian conversion

Testing done:
* Manual testing
* Added unit tests